+++ migrations/frontend/squashed.sql
@@ -4034,13 +4034,13 @@ CREATE TABLE product_subscriptions (
     archived_at timestamp with time zone,
     account_number text,
     cody_gateway_enabled boolean DEFAULT false NOT NULL,
-    cody_gateway_chat_rate_limit integer,
+    cody_gateway_chat_rate_limit bigint,
     cody_gateway_chat_rate_interval_seconds integer,
-    cody_gateway_embeddings_api_rate_limit integer,
+    cody_gateway_embeddings_api_rate_limit bigint,
     cody_gateway_embeddings_api_rate_interval_seconds integer,
     cody_gateway_embeddings_api_allowed_models text[],
     cody_gateway_chat_rate_limit_allowed_models text[],
-    cody_gateway_code_rate_limit integer,
+    cody_gateway_code_rate_limit bigint,
     cody_gateway_code_rate_interval_seconds integer,
     cody_gateway_code_rate_limit_allowed_models text[]
 );
